Extract stdlib utility helpers to doctest/parts/std/utility.h
