+++ doctest/doctest.h
@@ -612,6 +612,28 @@ namespace types {
 } // namespace types
 } // namespace detail
 } // namespace doctest
+namespace doctest {
+namespace detail {
+
+  // <utility>
+  template <typename T>
+  T&& declval();
+
+  template <class T>
+  DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type& t) DOCTEST_NOEXCEPT {
+      return static_cast<T&&>(t);
+  }
+
+  template <class T>
+  DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type&& t) DOCTEST_NOEXCEPT {
+      return static_cast<T&&>(t);
+  }
+
+  template <typename T>
+  struct deferred_false : types::false_type { };
+
+} // namespace detail
+} // namespace doctest
 
 namespace doctest {
 
@@ -996,23 +1018,6 @@ struct ContextOptions //!OCLINT too many fields
 };
 
 namespace detail {
-    // <utility>
-    template <typename T>
-    T&& declval();
-
-    template <class T>
-    DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type& t) DOCTEST_NOEXCEPT {
-        return static_cast<T&&>(t);
-    }
-
-    template <class T>
-    DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type&& t) DOCTEST_NOEXCEPT {
-        return static_cast<T&&>(t);
-    }
-
-    template <typename T>
-    struct deferred_false : types::false_type { };
-
 // MSVS 2015 :(
 #if !DOCTEST_CLANG && defined(_MSC_VER) && _MSC_VER <= 1900
     template <typename T, typename = void>
