+++ doctest/parts/doctest_fwd.h
@@ -63,6 +63,7 @@ DOCTEST_MSVC_SUPPRESS_WARNING(4623) // default constructor was implicitly define
 #include <doctest/parts/public/debugger.h>
 #include <doctest/parts/public/std/fwd.h>
 #include <doctest/parts/public/std/type_traits.h>
+#include <doctest/parts/public/std/utility.h>
 
 namespace doctest {
 
@@ -447,23 +448,6 @@ struct ContextOptions //!OCLINT too many fields
 };
 
 namespace detail {
-    // <utility>
-    template <typename T>
-    T&& declval();
-
-    template <class T>
-    DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type& t) DOCTEST_NOEXCEPT {
-        return static_cast<T&&>(t);
-    }
-
-    template <class T>
-    DOCTEST_CONSTEXPR_FUNC T&& forward(typename types::remove_reference<T>::type&& t) DOCTEST_NOEXCEPT {
-        return static_cast<T&&>(t);
-    }
-
-    template <typename T>
-    struct deferred_false : types::false_type { };
-
 // MSVS 2015 :(
 #if !DOCTEST_CLANG && defined(_MSC_VER) && _MSC_VER <= 1900
     template <typename T, typename = void>
